Move DeserializableClassScope and OverlayDeserializationRegistry into a new file.
