+++ BaseLib/CMakeLists.txt
@@ -17,6 +17,7 @@ SET( COMMON_SRCS
 	Log/unifiedLog.cpp
 	Log/debugLogger.cpp
 	uuid.cpp
+	Serialization/deserializableClassScope.cpp
 	Serialization/deserializer.cpp
 	Serialization/serializer.cpp
 	Serialization/serializerDeserializerCommon.cpp
