+++ BaseLib/Serialization/deserializationRegistry.cpp
@@ -24,21 +24,3 @@ babelwires::DeserializationRegistry::findEntry(std::string_view typeName) const 
     return nullptr;
 }
 
-babelwires::OverlayDeserializationRegistry::OverlayDeserializationRegistry(
-    const DeserializationRegistryInterface& baseRegistry)
-    : m_baseRegistry(baseRegistry) {}
-
-void babelwires::OverlayDeserializationRegistry::registerEntry(std::string_view typeName, const Entry* entry) {
-    assert(entry && "Cannot register a null deserialization entry");
-    const auto [it, wasInserted] = m_registeredEntries.insert(std::make_pair(std::string(typeName), entry));
-    assert((wasInserted || (it->second == entry)) && "The same type was registered with different entries");
-}
-
-const babelwires::DeserializationRegistryInterface::Entry*
-babelwires::OverlayDeserializationRegistry::findEntry(std::string_view typeName) const {
-    const auto it = m_registeredEntries.find(typeName);
-    if (it != m_registeredEntries.end()) {
-        return it->second;
-    }
-    return m_baseRegistry.findEntry(typeName);
-}
