+++ BaseLib/Serialization/deserializer.hpp
@@ -8,11 +8,11 @@
  **/
 #pragma once
 
-#include <BaseLib/Serialization/deserializationRegistry.hpp>
+#include <BaseLib/Result/result.hpp>
+#include <BaseLib/Serialization/deserializationRegistryInterface.hpp>
 #include <BaseLib/Serialization/serializable.hpp>
 #include <BaseLib/Serialization/serializableValue.hpp>
 #include <BaseLib/Serialization/serializerDeserializerCommon.hpp>
-#include <BaseLib/Result/result.hpp>
 
 #include <cassert>
 #include <cstdint>
@@ -70,13 +70,13 @@ namespace babelwires {
         /// Deserialize a child object of type T.
         /// Returns an error if the key is not found or if there is a parse error.
         /// The object is assigned by value, so any polymorphism is lost.
-        template<typename T>
+        template <typename T>
         babelwires::Result deserializeObjectByValue(T& object, std::string_view key = T::serializationType);
 
         /// Deserialize a child object of type T.
         /// If the key is not found, then false is returned and the object is not modified.
         /// The object is assigned by value, so any polymorphism is lost.
-        template<typename T>
+        template <typename T>
         babelwires::ResultT<bool> tryDeserializeObjectByValue(T& object, std::string_view key = T::serializationType);
 
         /// A non-standard iterator, which provides access to the deserialized objects of base type T in an array.
@@ -111,11 +111,12 @@ namespace babelwires {
 
         /// Get a description of the current parsing location which can be used in errors and warnings (e.g. a line
         /// number).
-        template <typename T>
-        void augmentResultWithContext(ResultT<T>& result) const;
+        template <typename T> void augmentResultWithContext(ResultT<T>& result) const;
 
+        /// Obtain the DeserializationRegistryInterface used by this Deserializer.
         const DeserializationRegistryInterface& getDeserializationRegistry() const;
 
+        /// Set the Deserializer to use the given DeserializationRegistryInterface.
         void setDeserializationRegistry(const DeserializationRegistryInterface& deserializationRegistry);
 
         /// Finalize the deserializer.
@@ -159,19 +160,6 @@ namespace babelwires {
         const DeserializationRegistryInterface* m_deserializationRegistry = nullptr;
     };
 
-    class DeserializableClassScope {
-      public:
-        explicit DeserializableClassScope(Deserializer& deserializer);
-        ~DeserializableClassScope();
-
-        template <typename T> void registerClass();
-
-      private:
-        Deserializer& m_deserializer;
-        const DeserializationRegistryInterface* m_previousRegistry = nullptr;
-        OverlayDeserializationRegistry m_overlayRegistry;
-    };
-
 } // namespace babelwires
 
 struct babelwires::Deserializer::BaseIterator {
@@ -206,10 +194,5 @@ template <typename T> struct babelwires::Deserializer::ValueIterator : BaseItera
 
     ValueIterator(std::unique_ptr<AbstractIterator> impl, Deserializer& deserializer, std::string_view typeName);
 };
-
-  template <typename T>
-  void babelwires::DeserializableClassScope::registerClass() {
-    m_overlayRegistry.registerClass<T>();
-  }
 
 #include <BaseLib/Serialization/deserializer_inl.hpp>
